Jest strict file renamed to satify filename rule.
