+++ src/config/index.js
@@ -1,5 +1,5 @@
 export {default as flowtype} from './flowtype';
-export {default as jestStrict} from './jest-strict';
+export {default as jestStrict} from './jest_strict';
 export {default as jest} from './jest';
 export {default as mocha} from './mocha';
 export {default as ops} from './ops';
