+++ src/config/jest_strict.js
